refactor: move state and logic into components, add auth context
