+++ monitor/src/components/layout/ProjectView.jsx
@@ -1,8 +1,8 @@
-import React from 'react'
+import React, { useState, useEffect, useRef, useCallback } from 'react'
 import { Card, CardHeader, CardTitle, CardContent } from '@/components/ui/card'
 import { Badge } from '@/components/ui/badge'
 import { Button } from '@/components/ui/button'
-import { Users, Sparkles, Settings, ScrollText, RefreshCw, Pause, Play, RotateCcw, Save, GitPullRequest, ArrowLeft, Github, Bell, ChevronDown, Lock, Unlock } from 'lucide-react'
+import { Users, User, Sparkles, Settings, ScrollText, RefreshCw, Pause, Play, RotateCcw, Save, GitPullRequest, ArrowLeft, Github, Bell, ChevronDown, Lock, Unlock } from 'lucide-react'
 import { PanelSlot } from '@/components/ui/panel'
 
 import Footer from '@/components/layout/Footer'
@@ -24,148 +24,556 @@ import BudgetInfoModal from '@/components/modals/BudgetInfoModal'
 import IntervalInfoModal from '@/components/modals/IntervalInfoModal'
 import TimeoutInfoModal from '@/components/modals/TimeoutInfoModal'
 import CreateIssueModal from '@/components/modals/CreateIssueModal'
+import { useAuth } from '@/hooks/useAuth'
+import { useNotifications } from '@/contexts/NotificationContext'
+import { useToast } from '@/contexts/ToastContext'
 
 export default function ProjectView({
   selectedProject,
-  goToProjectList,
+  setSelectedProject,
   projects,
   selectProject,
+  goToProjectList,
   error,
-  projectLoading,
   globalUptime,
-  isWriteMode,
-  handleLogout,
-  setLoginModal,
-  controlAction,
-  openBootstrapModal,
+  fetchGlobalStatus,
   notifCenter,
   setNotifCenter,
-  unreadCount,
-  repoUrl,
-  projectApi,
-  configForm,
-  configError,
-  configDirty,
-  configSaving,
-  updateConfigField,
-  resetConfig,
-  saveConfig,
-  setIntervalInfoModal,
-  setTimeoutInfoModal,
-  setBudgetInfoModal,
-  intervalInfoModal,
-  timeoutInfoModal,
-  budgetInfoModal,
-  agents,
-  prs,
-  comments,
-  commentsLoading,
-  loadMoreComments,
-  liveAgentLog,
-  setFocusedReportId,
-  setReportsPanelOpen,
-  issues,
-  issueFilter,
-  setIssueFilter,
-  openIssueModal,
-  setCreateIssueModal,
-  logs,
-  logsRef,
-  logsAutoFollow,
-  setLogsAutoFollow,
-  agentModal,
-  setAgentModal,
-  agentSettingsModal,
-  setAgentSettingsModal,
-  saveAgentSettings,
-  availableModels,
-  bootstrapModal,
-  setBootstrapModal,
-  executeBootstrap,
-  issueModal,
-  setIssueModal,
-  submitIssueComment,
-  reportsPanelOpen,
-  focusedReportId,
-  selectedAgent,
-  clearAgentFilter,
-  openAgentModal,
-  openAgentSettings,
-  createIssueModal,
-  createIssue,
-  selectAgent,
-  modKey,
-  notifList,
-  markAllRead,
-  markRead,
-  expandedNotifs,
-  toggleNotifExpand,
-  loginModal,
-  loginInput,
-  setLoginInput,
-  handleLogin,
-  projectSettingsOpen,
-  setProjectSettingsOpen,
-  settingsOpen,
-  setSettingsOpen,
   theme,
   setTheme,
-  notificationsEnabled,
-  toggleNotifications,
-  detailedNotifs,
-  setDetailedNotifs,
-  setShowApiKeyHelp,
-  globalTokenInput,
-  setGlobalTokenInput,
-  tokenSaving,
-  setTokenSaving,
-  setHasGlobalToken,
-  setGlobalTokenType,
-  setProviderTokens,
-  setGlobalTokenPreview,
-  setToast,
-  providerTokens,
-  codexLoginState,
-  setCodexLoginState,
-  authFetch,
-  setProjSetting,
-  notifUseGlobal,
-  projNotifSettings,
-  hasProjectToken,
-  projectTokenPreview,
-  projectTokenProviderLabel,
-  projectTokenSaving,
-  setProjectTokenSaving,
-  setHasProjectToken,
-  setProjectTokenPreview,
-  setProjectTokenProviderLabel,
-  projectTokenInput,
-  setProjectTokenInput,
-  projectTokenProvider,
-  setProjectTokenProvider,
-  projectCodexLoginState,
-  setProjectCodexLoginState,
-  config,
-  setSelectedProject,
-  fetchProjectData,
-  fetchGlobalStatus,
-  removeProject,
-  hasGlobalToken,
-  globalTokenPreview,
-  showApiKeyHelp,
-  toast,
+  onAgentChangeRef,
 }) {
+  const { isWriteMode, handleLogout, setLoginModal, loginModal, loginInput, setLoginInput, handleLogin, authFetch } = useAuth()
+  const { unreadCount } = useNotifications()
+  const { showToast, toast, setToast } = useToast()
+
+  // Project-specific state
+  const [logs, setLogs] = useState([])
+  const [agents, setAgents] = useState({ workers: [], managers: [] })
+  const [config, setConfig] = useState({ config: null, raw: '' })
+  const [prs, setPrs] = useState([])
+  const [issues, setIssues] = useState([])
+  const [issueFilter, setIssueFilter] = useState('open')
+  const [repoUrl, setRepoUrl] = useState(null)
+  const [projectLoading, setProjectLoading] = useState(false)
+  const [logsAutoFollow, setLogsAutoFollow] = useState(true)
+  const logsRef = useRef(null)
+
+  // Config state
+  const [configForm, setConfigForm] = useState({
+    cycleIntervalMs: 1800000, agentTimeoutMs: 900000, trackerIssue: 1, budgetPer24h: 0
+  })
+  const [configDirty, setConfigDirty] = useState(false)
+  const configDirtyRef = useRef(false)
+  const [configError, setConfigError] = useState(null)
+  const [configSaving, setConfigSaving] = useState(false)
+
+  // Reports state
+  const [comments, setComments] = useState([])
+  const [commentsPage, setCommentsPage] = useState(1)
+  const [commentsHasMore, setCommentsHasMore] = useState(true)
+  const [commentsLoading, setCommentsLoading] = useState(false)
+  const [selectedAgent, setSelectedAgent] = useState(() => localStorage.getItem('selectedAgent') || null)
+  const [reportsPanelOpen, setReportsPanelOpen] = useState(false)
+  const [focusedReportId, setFocusedReportId] = useState(null)
+  const [liveAgentLog, setLiveAgentLog] = useState(null)
+
+  // Scroll to focused report when panel opens
+  useEffect(() => {
+    if (reportsPanelOpen && focusedReportId) {
+      const timer = setTimeout(() => {
+        const el = document.querySelector(`[data-report-id="${focusedReportId}"]`)
+        if (el) el.scrollIntoView({ behavior: 'smooth', block: 'center' })
+        const clearTimer = setTimeout(() => setFocusedReportId(null), 2000)
+        return () => clearTimeout(clearTimer)
+      }, 350)
+      return () => clearTimeout(timer)
+    }
+  }, [reportsPanelOpen, focusedReportId])
+
+  // Modals
+  const [agentModal, setAgentModal] = useState({ open: false, agent: null, data: null, loading: false, tab: 'skill' })
+  const [issueModal, setIssueModal] = useState({ open: false, issue: null, comments: [], loading: false })
+  const [createIssueModal, setCreateIssueModal] = useState({ open: false, title: '', body: '', receiver: '', creating: false, error: null, focusedField: 'title' })
+  const modKey = typeof navigator !== 'undefined' && /Mac|iPhone|iPad/.test(navigator.userAgent) ? '⌘' : 'Ctrl'
+  const [bootstrapModal, setBootstrapModal] = useState({ open: false, loading: false, preview: null, error: null, executing: false, removeRoadmap: true, specMode: 'keep', specContent: '', whatToBuild: '', successCriteria: '' })
+  const [budgetInfoModal, setBudgetInfoModal] = useState(false)
+  const [intervalInfoModal, setIntervalInfoModal] = useState(false)
+  const [timeoutInfoModal, setTimeoutInfoModal] = useState(false)
+  const [settingsOpen, setSettingsOpen] = useState(false)
+  const [showApiKeyHelp, setShowApiKeyHelp] = useState(false)
+  const [projectSettingsOpen, setProjectSettingsOpen] = useState(false)
+
+  // Project settings
+  const [hasProjectToken, setHasProjectToken] = useState(false)
+  const [projectTokenPreview, setProjectTokenPreview] = useState(null)
+  const [projectTokenProviderLabel, setProjectTokenProviderLabel] = useState(null)
+  const [projectTokenInput, setProjectTokenInput] = useState('')
+  const [projectTokenProvider, setProjectTokenProvider] = useState('')
+  const [projectTokenSaving, setProjectTokenSaving] = useState(false)
+  const [projectCodexLoginState, setProjectCodexLoginState] = useState(null)
+  const [codexLoginState, setCodexLoginState] = useState(null)
+  const [hasGlobalToken, setHasGlobalToken] = useState(false)
+  const [globalTokenPreview, setGlobalTokenPreview] = useState(null)
+
+  // Agent settings modal
+  const [agentSettingsModal, setAgentSettingsModal] = useState({ open: false, agent: null, model: '', saving: false, error: null })
+  const [availableModels, setAvailableModels] = useState([])
+
+  // Per-project notification settings
+  const [projectNotifs, setProjectNotifs] = useState(() => {
+    try { return JSON.parse(localStorage.getItem('tbc_project_notifs') || '{}') } catch { return {} }
+  })
+
+  const selectedProjectRef = useRef(null)
+  useEffect(() => { selectedProjectRef.current = selectedProject }, [selectedProject])
+
+  const projectApi = (path) => selectedProject ? `/api/projects/${selectedProject.id}${path}` : null
+
+  // Fetch project data
+  const fetchProjectData = useCallback(async (initial = false) => {
+    const currentProject = selectedProjectRef.current
+    const baseApi = currentProject ? `/api/projects/${currentProject.id}` : null
+    if (!baseApi) return
+    if (initial) setProjectLoading(true)
+    
+    try {
+      const [logsRes, agentsRes, configRes, prsRes, issuesRes, repoRes] = await Promise.all([
+        fetch(`${baseApi}/logs?lines=100`),
+        fetch(`${baseApi}/agents`),
+        fetch(`${baseApi}/config`),
+        fetch(`${baseApi}/prs`),
+        fetch(`${baseApi}/issues`).catch(() => ({ ok: false })),
+        fetch(`${baseApi}/repo`)
+      ])
+      
+      if (selectedProjectRef.current?.id !== currentProject.id) return
+      
+      setLogs((await logsRes.json()).logs || [])
+      setAgents(await agentsRes.json())
+      
+      const configData = await configRes.json()
+      setConfig(configData)
+      setHasProjectToken(!!configData.hasProjectToken)
+      setProjectTokenPreview(configData.projectTokenPreview || null)
+      setProjectTokenProviderLabel(configData.provider || null)
+      if (!configDirtyRef.current && configData.config) {
+        setConfigForm({
+          cycleIntervalMs: configData.config.cycleIntervalMs ?? 1800000,
+          agentTimeoutMs: configData.config.agentTimeoutMs ?? 900000,
+          trackerIssue: configData.config.trackerIssue ?? 1,
+          budgetPer24h: configData.config.budgetPer24h ?? 0
+        })
+      }
+      
+      setPrs((await prsRes.json()).prs || [])
+      if (issuesRes.ok) {
+        setIssues((await issuesRes.json()).issues || [])
+      }
+      setRepoUrl((await repoRes.json()).url)
+    } catch (err) {
+      console.error('Failed to fetch project data:', err)
+      if (initial) showToast('Failed to load project data')
+    } finally {
+      setProjectLoading(false)
+    }
+  }, [showToast])
+
+  const fetchComments = useCallback(async (page = 1, agent = null, append = false, silent = false) => {
+    const currentProject = selectedProjectRef.current
+    const baseApi = currentProject ? `/api/projects/${currentProject.id}` : null
+    if (!baseApi) return
+    if (!silent) setCommentsLoading(true)
+    try {
+      const params = new URLSearchParams({ page, per_page: 10 })
+      if (agent) params.set('agent', agent)
+      const res = await fetch(`${baseApi}/reports?${params}`)
+      if (!res.ok) return
+      if (selectedProjectRef.current?.id !== currentProject.id) return
+      const data = await res.json()
+      if (append) setComments(prev => [...prev, ...data.reports])
+      else setComments(data.reports || [])
+      setCommentsHasMore((data.page * data.perPage) < data.total)
+      setCommentsPage(page)
+    } catch (err) { console.error('Failed to fetch comments:', err) }
+    finally { if (!silent) setCommentsLoading(false) }
+  }, [])
+
+  // Register agent change callback
+  useEffect(() => {
+    if (onAgentChangeRef) {
+      onAgentChangeRef.current = () => {
+        fetchProjectData()
+        fetchComments(1, localStorage.getItem('selectedAgent') || null, false)
+      }
+    }
+    return () => { if (onAgentChangeRef) onAgentChangeRef.current = null }
+  }, [fetchProjectData, fetchComments, onAgentChangeRef])
+
+  // Fetch on project change
+  useEffect(() => {
+    if (selectedProject) {
+      // Reset state for new project
+      setLogs([])
+      setAgents({ workers: [], managers: [] })
+      setComments([])
+      setCommentsPage(1)
+      setPrs([])
+      setIssues([])
+      setIssueFilter('open')
+
+      fetchProjectData(true)
+      const savedAgent = localStorage.getItem('selectedAgent')
+      fetchComments(1, savedAgent, false)
+      
+      const fetchLogs = async () => {
+        const api = `/api/projects/${selectedProject.id}/logs?lines=100`
+        try {
+          const res = await fetch(api)
+          if (!res.ok) return
+          setLogs((await res.json()).logs || [])
+        } catch {}
+      }
+
+      const logsInterval = setInterval(fetchLogs, 10000)
+      const commentsInterval = setInterval(() => fetchComments(1, localStorage.getItem('selectedAgent') || null, false, true), 30000)
+      const projectDataInterval = setInterval(fetchProjectData, 30000)
+      
+      return () => {
+        clearInterval(logsInterval)
+        clearInterval(commentsInterval)
+        clearInterval(projectDataInterval)
+      }
+    }
+  }, [selectedProject?.id])
+
+  // Fetch settings & models on mount
+  useEffect(() => {
+    fetch('/api/settings').then(r => r.json()).then(d => {
+      setHasGlobalToken(!!d.hasGlobalToken)
+      setGlobalTokenPreview(d.globalTokenPreview || null)
+    }).catch(() => {})
+    fetch('/api/models').then(r => r.json()).then(data => {
+      if (data.data) setAvailableModels(data.data)
+    }).catch(() => {})
+    fetch('/api/openai-codex/status').then(r => r.json()).then(d => {
+      if (d.authenticated) setCodexLoginState('success')
+    }).catch(() => {})
+  }, [])
+
+  // Check project-level codex auth when project changes
+  useEffect(() => {
+    if (selectedProject?.id) {
+      fetch(`/api/openai-codex/status?project=${encodeURIComponent(selectedProject.id)}`).then(r => r.json()).then(d => {
+        setProjectCodexLoginState(d.authenticated ? 'success' : null)
+      }).catch(() => setProjectCodexLoginState(null))
+    } else {
+      setProjectCodexLoginState(null)
+    }
+  }, [selectedProject?.id])
+
+  // Poll for live agent log
+  useEffect(() => {
+    if (!selectedProject?.currentAgent) {
+      setLiveAgentLog(null)
+      return
+    }
+    const fetchAgentLog = async () => {
+      try {
+        const res = await fetch(`/api/projects/${selectedProject.id}/agent-log`)
+        if (!res.ok) return
+        const data = await res.json()
+        if (data.running) {
+          setLiveAgentLog({ agent: data.agent, model: data.model, startTime: data.startTime, log: data.log })
+        } else {
+          setLiveAgentLog(null)
+        }
+      } catch {}
+    }
+    fetchAgentLog()
+    const interval = setInterval(fetchAgentLog, 3000)
+    return () => clearInterval(interval)
+  }, [selectedProject?.id, selectedProject?.currentAgent])
+
+  // Auto-scroll logs
+  useEffect(() => {
+    if (logsAutoFollow && logsRef.current) {
+      logsRef.current.scrollTop = logsRef.current.scrollHeight
+    }
+  }, [logs, logsAutoFollow])
+
+  // Actions
+  const controlAction = async (action) => {
+    if (!selectedProject) return
+    try {
+      const res = await authFetch(projectApi(`/${action}`), { method: 'POST' })
+      if (res.ok) await fetchGlobalStatus()
+      else showToast(`Action "${action}" failed`)
+    } catch (err) { showToast(`Action "${action}" failed: ${err.message}`) }
+  }
+
+  const saveConfig = async () => {
+    if (!selectedProject) return
+    setConfigSaving(true)
+    setConfigError(null)
+    try {
+      const existing = config.config || {}
+      const merged = { ...existing,
+        cycleIntervalMs: configForm.cycleIntervalMs,
+        agentTimeoutMs: configForm.agentTimeoutMs,
+        trackerIssue: configForm.trackerIssue,
+      }
+      if (configForm.budgetPer24h > 0) merged.budgetPer24h = configForm.budgetPer24h
+      else delete merged.budgetPer24h
+      const lines = [`# ${selectedProject.id} - Orchestrator Configuration`]
+      const scalarKeys = Object.keys(merged).filter(k => typeof merged[k] !== 'object' || merged[k] === null)
+      const objectKeys = Object.keys(merged).filter(k => typeof merged[k] === 'object' && merged[k] !== null)
+      for (const k of scalarKeys) lines.push(`${k}: ${merged[k]}`)
+      for (const k of objectKeys) {
+        lines.push(`${k}:`)
+        for (const [sk, sv] of Object.entries(merged[k])) {
+          if (typeof sv === 'object' && sv !== null) {
+            lines.push(`  ${sk}:`)
+            for (const [ssk, ssv] of Object.entries(sv)) lines.push(`    ${ssk}: ${ssv}`)
+          } else {
+            lines.push(`  ${sk}: ${sv}`)
+          }
+        }
+      }
+      const yaml = lines.join('\n') + '\n'
+      const res = await authFetch(projectApi('/config'), {
+        method: 'POST', headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify({ content: yaml })
+      })
+      const data = await res.json()
+      if (!res.ok) throw new Error(data.error || 'Failed to save')
+      configDirtyRef.current = false
+      setConfigDirty(false)
+      await fetchProjectData()
+    } catch (err) { setConfigError(err.message) }
+    finally { setConfigSaving(false) }
+  }
+
+  const updateConfigField = (field, value) => {
+    configDirtyRef.current = true
+    setConfigForm(prev => ({ ...prev, [field]: value }))
+    setConfigDirty(true)
+    setConfigError(null)
+  }
+
+  const resetConfig = () => {
+    if (config.config) {
+      setConfigForm({
+        cycleIntervalMs: config.config.cycleIntervalMs ?? 1800000,
+        agentTimeoutMs: config.config.agentTimeoutMs ?? 900000,
+        trackerIssue: config.config.trackerIssue ?? 1,
+        budgetPer24h: config.config.budgetPer24h ?? 0
+      })
+    }
+    configDirtyRef.current = false
+    setConfigDirty(false)
+    setConfigError(null)
+  }
+
+  const loadMoreComments = () => {
+    if (!commentsLoading && commentsHasMore) fetchComments(commentsPage + 1, selectedAgent, true)
+  }
+
+  const selectAgentFilter = (agent) => {
+    setSelectedAgent(agent)
+    localStorage.setItem('selectedAgent', agent)
+    setCommentsPage(1)
+    fetchComments(1, agent, false)
+  }
+
+  const clearAgentFilter = () => {
+    setSelectedAgent(null)
+    localStorage.removeItem('selectedAgent')
+    setCommentsPage(1)
+    fetchComments(1, null, false)
+  }
+
+  const openAgentModal = async (agentName) => {
+    if (!selectedProject) return
+    setAgentModal({ open: true, agent: agentName, data: null, loading: true, tab: 'skill' })
+    try {
+      const res = await fetch(projectApi(`/agents/${agentName}`))
+      const data = await res.json()
+      setAgentModal({ open: true, agent: agentName, data, loading: false, tab: 'skill' })
+    } catch {
+      setAgentModal({ open: true, agent: agentName, data: null, loading: false, tab: 'skill' })
+    }
+  }
+
+  const openAgentSettings = (agent) => {
+    setAgentSettingsModal({ open: false, agent: null, model: '', saving: false, error: null })
+    fetch(projectApi(`/agents/${agent.name}`))
+      .then(r => r.json())
+      .then(data => {
+        setAgentSettingsModal({ open: true, agent, model: data.model || '', saving: false, error: null })
+      })
+      .catch(() => {
+        setAgentSettingsModal({ open: true, agent, model: '', saving: false, error: null })
+      })
+  }
+
+  const saveAgentSettings = async () => {
+    setAgentSettingsModal(prev => ({ ...prev, saving: true, error: null }))
+    try {
+      const res = await fetch(projectApi(`/agents/${agentSettingsModal.agent.name}`), {
+        method: 'PATCH',
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify({ model: agentSettingsModal.model })
+      })
+      if (!res.ok) throw new Error((await res.json()).error || 'Failed to save')
+      setAgentSettingsModal(prev => ({ ...prev, open: false }))
+      fetchProjectData()
+    } catch (e) {
+      setAgentSettingsModal(prev => ({ ...prev, saving: false, error: e.message }))
+    }
+  }
+
+  const openBootstrapModal = async () => {
+    if (!selectedProject) return
+    setBootstrapModal({ open: true, loading: true, preview: null, error: null, executing: false, removeRoadmap: true, specMode: 'keep', specContent: '', whatToBuild: '', successCriteria: '' })
+    try {
+      const res = await authFetch(projectApi('/bootstrap'))
+      const data = await res.json()
+      setBootstrapModal({ open: true, loading: false, preview: data, error: null, executing: false, removeRoadmap: !!data.hasRoadmap, specMode: 'keep', specContent: data.specContent || '', whatToBuild: '', successCriteria: '' })
+    } catch (err) {
+      setBootstrapModal({ open: true, loading: false, preview: null, error: err.message, executing: false, removeRoadmap: true, specMode: 'keep', specContent: '', whatToBuild: '', successCriteria: '' })
+    }
+  }
+
+  const executeBootstrap = async () => {
+    if (!selectedProject) return
+    setBootstrapModal(prev => ({ ...prev, executing: true, error: null }))
+    try {
+      const body = {
+        removeRoadmap: bootstrapModal.removeRoadmap,
+        spec: {
+          mode: bootstrapModal.specMode,
+          content: bootstrapModal.specContent,
+          whatToBuild: bootstrapModal.whatToBuild,
+          successCriteria: bootstrapModal.successCriteria,
+        }
+      }
+      const res = await authFetch(projectApi('/bootstrap'), {
+        method: 'POST',
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify(body)
+      })
+      const data = await res.json()
+      if (!res.ok) throw new Error(data.error)
+      setBootstrapModal({ open: false, loading: false, preview: null, error: null, executing: false, removeRoadmap: true, specMode: 'keep', specContent: '', whatToBuild: '', successCriteria: '' })
+      await fetchGlobalStatus()
+      await fetchProjectData()
+      fetchComments(1, selectedAgent, false)
+    } catch (err) {
+      setBootstrapModal(prev => ({ ...prev, executing: false, error: err.message }))
+    }
+  }
+
+  const createIssue = async () => {
+    if (!createIssueModal.title.trim()) return
+    setCreateIssueModal(prev => ({ ...prev, creating: true, error: null }))
+    try {
+      const res = await authFetch(projectApi('/issues/create'), {
+        method: 'POST',
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify({
+          title: createIssueModal.title.trim(),
+          body: createIssueModal.body.trim(),
+          creator: 'human',
+          assignee: createIssueModal.receiver || null
+        })
+      })
+      const data = await res.json()
+      if (data.success) {
+        setCreateIssueModal({ open: false, title: '', body: '', receiver: '', creating: false, error: null, focusedField: 'title' })
+        await fetchProjectData()
+      } else {
+        setCreateIssueModal(prev => ({ ...prev, creating: false, error: data.error || 'Failed to create issue' }))
+      }
+    } catch (err) {
+      setCreateIssueModal(prev => ({ ...prev, creating: false, error: err.message }))
+    }
+  }
+
+  const openIssueModal = async (issueId) => {
+    if (!selectedProject) return
+    setIssueModal({ open: true, issue: null, comments: [], loading: true })
+    try {
+      const res = await fetch(projectApi(`/issues/${issueId}`))
+      const data = await res.json()
+      setIssueModal({ open: true, issue: data.issue, comments: data.comments || [], loading: false })
+    } catch {
+      setIssueModal({ open: true, issue: null, comments: [], loading: false })
+    }
+  }
+
+  const submitIssueComment = async () => {
+    if (!issueModal.issue || !issueModal.newComment?.trim()) return
+    setIssueModal(prev => ({ ...prev, commenting: true }))
+    try {
+      const res = await authFetch(projectApi(`/issues/${issueModal.issue.id}/comments`), {
+        method: 'POST',
+        headers: { 'Content-Type': 'application/json' },
+        body: JSON.stringify({ author: 'human', body: issueModal.newComment.trim() })
+      })
+      if (res.ok) {
+        const data = await (await fetch(projectApi(`/issues/${issueModal.issue.id}`))).json()
+        setIssueModal(prev => ({ ...prev, issue: data.issue, comments: data.comments || [], newComment: '', commenting: false }))
+      } else {
+        setIssueModal(prev => ({ ...prev, commenting: false }))
+      }
+    } catch {
+      setIssueModal(prev => ({ ...prev, commenting: false }))
+    }
+  }
+
+  const removeProject = async (projectId) => {
+    try {
+      const res = await authFetch(`/api/projects/${projectId}`, { method: 'DELETE' })
+      if (res.ok) {
+        if (selectedProject?.id === projectId) {
+          goToProjectList()
+        }
+        await fetchGlobalStatus()
+      }
+    } catch (err) {
+      console.error('Failed to remove project:', err)
+    }
+  }
+
+  // Project notification settings helpers
+  const getProjSetting = (section) => {
+    if (!selectedProject) return { useGlobal: true }
+    const all = projectNotifs[selectedProject.id] || {}
+    return all[section] || { useGlobal: true }
+  }
+  const setProjSetting = (section, patch) => {
+    if (!selectedProject) return
+    const all = { ...projectNotifs }
+    const proj = { ...(all[selectedProject.id] || {}) }
+    proj[section] = { ...(proj[section] || { useGlobal: true }), ...patch }
+    all[selectedProject.id] = proj
+    setProjectNotifs(all)
+    localStorage.setItem('tbc_project_notifs', JSON.stringify(all))
+  }
+
+  const projNotifSettings = getProjSetting('notifs')
+  const notifUseGlobal = projNotifSettings.useGlobal !== false
+
   if (!selectedProject) return null
 
   return (
     <div className="flex min-h-screen">
     <div className="flex-1 min-w-0 bg-neutral-50 dark:bg-neutral-950 p-6">
       <div>
-        {/* Header - Mobile Friendly */}
+        {/* Header */}
         <div className="mb-6 space-y-3">
-          {/* Desktop: single row. Mobile: two rows */}
           <div className="flex flex-col sm:flex-row sm:items-center sm:justify-between gap-3">
-            {/* Left: Back button + Title */}
             <div className="flex items-center gap-2">
               <Button variant="ghost" size="sm" onClick={goToProjectList} className="text-neutral-500 dark:text-neutral-400 shrink-0 px-2">
                 <ArrowLeft className="w-4 h-4" />
@@ -174,7 +582,6 @@ export default function ProjectView({
               <h1 className="text-lg sm:text-2xl font-bold text-neutral-800 dark:text-neutral-100 truncate">{selectedProject.id}</h1>
             </div>
             
-            {/* Right: Actions */}
             <div className="flex items-center gap-1.5 pl-8 sm:pl-0 shrink-0">
               <button
                 onClick={() => setNotifCenter(true)}
@@ -225,7 +632,6 @@ export default function ProjectView({
             </div>
           </div>
           
-          {/* Project tabs - hidden on mobile */}
           {projects.length > 1 && (
             <div className="hidden sm:flex items-center gap-1.5 overflow-x-auto pb-1">
               {projects.filter(p => !p.archived || p.id === selectedProject?.id).map(project => (
@@ -259,7 +665,6 @@ export default function ProjectView({
             {/* Phase Indicator */}
             {selectedProject.phase && (
               <div className="mb-4 p-4 rounded-lg bg-white dark:bg-neutral-900 border border-neutral-200 dark:border-neutral-700">
-                {/* Row 1: Phase badge */}
                 <div className="flex items-center gap-2 mb-2">
                   <span className="text-lg">
                     {selectedProject.phase === 'athena' ? '🧠' : selectedProject.phase === 'implementation' ? '🔨' : selectedProject.phase === 'verification' ? '✅' : '❓'}
@@ -277,7 +682,6 @@ export default function ProjectView({
                     </span>
                   )}
                 </div>
-                {/* Progress bar */}
                 {selectedProject.milestoneCyclesBudget > 0 && selectedProject.phase === 'implementation' && (
                   <div className="w-full h-1.5 bg-neutral-200 dark:bg-neutral-700 rounded-full overflow-hidden mb-2">
                     <div
@@ -289,7 +693,6 @@ export default function ProjectView({
                     />
                   </div>
                 )}
-                {/* Row 2: Milestone description */}
                 {(selectedProject.milestoneTitle || selectedProject.milestone) && (
                   <details className="group">
                     <summary className="text-sm text-neutral-700 dark:text-neutral-300 cursor-pointer hover:text-neutral-900 dark:hover:text-neutral-100 flex items-start gap-1.5 [&::-webkit-details-marker]:hidden list-none">
@@ -304,7 +707,6 @@ export default function ProjectView({
 
             {/* Row 1: State, Cost & Budget, Config */}
             <div className="grid gap-4" style={{ gridTemplateColumns: "repeat(auto-fit, minmax(300px, 1fr))" }}>
-              {/* State */}
               <OrchestratorStateCard
                 selectedProject={selectedProject}
                 globalUptime={globalUptime}
@@ -312,13 +714,11 @@ export default function ProjectView({
                 isWriteMode={isWriteMode}
               />
 
-              {/* Cost & Budget */}
               <CostBudgetCard
                 selectedProject={selectedProject}
                 setBudgetInfoModal={setBudgetInfoModal}
               />
 
-              {/* Config */}
               <ConfigCard
                 configForm={configForm}
                 configError={configError}
@@ -333,7 +733,6 @@ export default function ProjectView({
                 setBudgetInfoModal={setBudgetInfoModal}
               />
 
-              {/* Managers */}
               <Card>
                 <CardHeader><CardTitle className="flex items-center gap-2"><Sparkles className="w-4 h-4" />Managers ({agents.managers.length})</CardTitle></CardHeader>
                 <CardContent>
@@ -347,7 +746,7 @@ export default function ProjectView({
                         selectedAgent={selectedAgent}
                         openAgentModal={openAgentModal}
                         openAgentSettings={openAgentSettings}
-                        selectAgent={selectAgent}
+                        selectAgent={selectAgentFilter}
                         clearAgentFilter={clearAgentFilter}
                       />
                     ))}
@@ -356,7 +755,6 @@ export default function ProjectView({
                 </CardContent>
               </Card>
 
-              {/* Workers */}
               <Card>
                 <CardHeader><CardTitle className="flex items-center gap-2"><Users className="w-4 h-4" />Workers ({agents.workers.length})</CardTitle></CardHeader>
                 <CardContent>
@@ -369,7 +767,7 @@ export default function ProjectView({
                         selectedAgent={selectedAgent}
                         openAgentModal={openAgentModal}
                         openAgentSettings={openAgentSettings}
-                        selectAgent={selectAgent}
+                        selectAgent={selectAgentFilter}
                         clearAgentFilter={clearAgentFilter}
                       />
                     ))}
@@ -378,7 +776,6 @@ export default function ProjectView({
                 </CardContent>
               </Card>
 
-              {/* PRs */}
               <Card>
                 <CardHeader><CardTitle className="flex items-center gap-2"><GitPullRequest className="w-4 h-4" />Open PRs ({prs.length})</CardTitle></CardHeader>
                 <CardContent>
@@ -401,7 +798,6 @@ export default function ProjectView({
                 </CardContent>
               </Card>
 
-              {/* Agent Reports */}
               <AgentReportsCard
                 comments={comments}
                 commentsLoading={commentsLoading}
@@ -412,7 +808,6 @@ export default function ProjectView({
                 setReportsPanelOpen={setReportsPanelOpen}
               />
 
-              {/* Issues */}
               <IssuesSidebar
                 issues={issues}
                 issueFilter={issueFilter}
@@ -423,7 +818,7 @@ export default function ProjectView({
               />
             </div>
 
-            {/* Row 4: Logs */}
+            {/* Logs */}
             <Card className="mt-4">
               <CardHeader><CardTitle className="flex items-center gap-2"><ScrollText className="w-4 h-4" />Orchestrator Logs</CardTitle></CardHeader>
               <CardContent>
@@ -449,35 +844,14 @@ export default function ProjectView({
         )}
       </div>
 
-      {/* Agent Details Panel */}
       <AgentDetailPanel agentModal={agentModal} setAgentModal={setAgentModal} />
-
-      {/* Agent Settings Modal */}
       <AgentSettingsModal agentSettingsModal={agentSettingsModal} setAgentSettingsModal={setAgentSettingsModal} saveAgentSettings={saveAgentSettings} />
-
-      {/* Bootstrap Panel */}
-      <BootstrapPanel
-        bootstrapModal={bootstrapModal}
-        setBootstrapModal={setBootstrapModal}
-        executeBootstrap={executeBootstrap}
-      />
-
-      {/* Budget Info Modal */}
+      <BootstrapPanel bootstrapModal={bootstrapModal} setBootstrapModal={setBootstrapModal} executeBootstrap={executeBootstrap} />
       <BudgetInfoModal open={budgetInfoModal} onClose={() => setBudgetInfoModal(false)} />
-
-      {/* Interval Info Modal */}
       <IntervalInfoModal open={intervalInfoModal} onClose={() => setIntervalInfoModal(false)} />
-
-      {/* Timeout Info Modal */}
       <TimeoutInfoModal open={timeoutInfoModal} onClose={() => setTimeoutInfoModal(false)} />
-
-      {/* API Key Help Modal */}
       <ApiKeyHelpModal open={showApiKeyHelp} onClose={() => setShowApiKeyHelp(false)} />
-
-      {/* Create Issue Modal */}
       <CreateIssueModal createIssueModal={createIssueModal} setCreateIssueModal={setCreateIssueModal} createIssue={createIssue} agents={agents} modKey={modKey} />
-
-      {/* Issue Detail Panel */}
       <IssueDetailPanel
         issueModal={issueModal}
         setIssueModal={setIssueModal}
@@ -487,8 +861,6 @@ export default function ProjectView({
         submitIssueComment={submitIssueComment}
         modKey={modKey}
       />
-
-      {/* Agent Reports Panel */}
       <ReportsPanel
         open={reportsPanelOpen}
         onClose={() => setReportsPanelOpen(false)}
@@ -502,8 +874,6 @@ export default function ProjectView({
         clearAgentFilter={clearAgentFilter}
         selectedProject={selectedProject}
       />
-
-      {/* Login Modal */}
       <LoginModal
         open={loginModal}
         onClose={() => { setLoginModal(false); setLoginInput('') }}
@@ -511,32 +881,13 @@ export default function ProjectView({
         setLoginInput={setLoginInput}
         handleLogin={handleLogin}
       />
-
       <SettingsPanel
         settingsOpen={settingsOpen}
         onClose={() => setSettingsOpen(false)}
         theme={theme}
         setTheme={setTheme}
-        notificationsEnabled={notificationsEnabled}
-        toggleNotifications={toggleNotifications}
-        detailedNotifs={detailedNotifs}
-        setDetailedNotifs={setDetailedNotifs}
         setShowApiKeyHelp={setShowApiKeyHelp}
-        globalTokenInput={globalTokenInput}
-        setGlobalTokenInput={setGlobalTokenInput}
-        tokenSaving={tokenSaving}
-        setTokenSaving={setTokenSaving}
-        setHasGlobalToken={setHasGlobalToken}
-        setGlobalTokenType={setGlobalTokenType}
-        setProviderTokens={setProviderTokens}
-        setGlobalTokenPreview={setGlobalTokenPreview}
-        setToast={setToast}
-        providerTokens={providerTokens}
-        codexLoginState={codexLoginState}
-        setCodexLoginState={setCodexLoginState}
-        authFetch={authFetch}
       />
-      {/* Project Settings Panel */}
       <ProjectSettingsPanel
         selectedProject={selectedProject}
         projectSettingsOpen={projectSettingsOpen}
@@ -555,7 +906,7 @@ export default function ProjectView({
         setHasProjectToken={setHasProjectToken}
         setProjectTokenPreview={setProjectTokenPreview}
         setProjectTokenProviderLabel={setProjectTokenProviderLabel}
-        setToast={setToast}
+        setToast={showToast}
         projectTokenInput={projectTokenInput}
         setProjectTokenInput={setProjectTokenInput}
         projectTokenProvider={projectTokenProvider}
@@ -572,16 +923,9 @@ export default function ProjectView({
         hasGlobalToken={hasGlobalToken}
         globalTokenPreview={globalTokenPreview}
       />
-      {/* Notification Center */}
       <NotificationPanel
         open={notifCenter}
         onClose={() => setNotifCenter(false)}
-        notifList={notifList}
-        unreadCount={unreadCount}
-        markAllRead={markAllRead}
-        markRead={markRead}
-        expandedNotifs={expandedNotifs}
-        toggleNotifExpand={toggleNotifExpand}
       />
 
       {/* Toast notifications */}
