+++ monitor/src/contexts/ToastContext.jsx
@@ -1,0 +1,36 @@
+import React, { createContext, useState, useCallback, useContext } from 'react'
+
+const ToastContext = createContext(null)
+
+export function ToastProvider({ children }) {
+  const [toast, setToast] = useState(null)
+
+  const showToast = useCallback((message, type = 'error') => {
+    setToast({ message, type })
+    setTimeout(() => setToast(null), 4000)
+  }, [])
+
+  return (
+    <ToastContext.Provider value={{ toast, setToast, showToast }}>
+      {children}
+      {toast && (
+        <div className={`fixed bottom-4 right-4 z-50 px-4 py-3 rounded-lg shadow-lg text-sm font-medium transition-all ${
+          (typeof toast === 'object' ? toast.type : '') === 'error' ? 'bg-red-600 text-white' :
+          (typeof toast === 'object' ? toast.type : '') === 'success' ? 'bg-green-600 text-white' :
+          'bg-neutral-800 text-white'
+        }`}>
+          <div className="flex items-center gap-2">
+            <span>{typeof toast === 'string' ? toast : toast.message}</span>
+            <button onClick={() => setToast(null)} className="ml-2 opacity-70 hover:opacity-100">✕</button>
+          </div>
+        </div>
+      )}
+    </ToastContext.Provider>
+  )
+}
+
+export function useToast() {
+  const ctx = useContext(ToastContext)
+  if (!ctx) throw new Error('useToast must be used within ToastProvider')
+  return ctx
+}
