+++ monitor/src/hooks/useAuth.js
@@ -1,0 +1,8 @@
+import { useContext } from 'react'
+import { AuthContext } from '@/contexts/AuthContext'
+
+export function useAuth() {
+  const ctx = useContext(AuthContext)
+  if (!ctx) throw new Error('useAuth must be used within AuthProvider')
+  return ctx
+}
